Added function to delete data from post
Update PostController and index view to add confirmation before deleting a post

The PostController's destroy method has been updated to use a database transaction to detach tags and categories before deleting the post. If an exception is thrown, the transaction is rolled back and an error alert is displayed. If the deletion is successful, a success alert is displayed and the user is redirected to the posts index page.

The index view has been updated to include a confirmation step before submitting the delete form. When the form is submitted, a SweetAlert2 dialog is displayed asking the user to confirm the deletion. If the user confirms, the form is submitted. If the user cancels, the form submission is cancelled.

A new internal JavaScript section has been added to the index view to handle the delete form submission and display the SweetAlert2 dialog.