"Paginate posts and add pagination view in index page"
This commit changes the way posts are displayed in the index page of the application. Instead of showing all posts at once, the posts are now paginated with 10 posts per page. Additionally, a pagination view has been added to the bottom of the index page, allowing the user to navigate between pages.