Displaying user data
The UserController's index method has been updated to pass all users to the view using the `User::all()` method. Additionally, the `users.index` view has been modified to display a list of users using a `@forelse` loop. Each user is displayed in a separate card containing their name, email, and role. A delete button and an edit button are also included for each user. If no users are found, a message is displayed indicating that there is no data to fetch.